+++ MantidPlot/src/Mantid/MantidPlotUtilities.h
@@ -10,11 +10,11 @@
 
 /// Returns a single log value from the given workspace
 double getSingleWorkspaceLogValue(
-    int wsIndex, const Mantid::API::MatrixWorkspace_const_sptr &matrixWS,
+    size_t wsIndex, const Mantid::API::MatrixWorkspace_const_sptr &matrixWS,
     const QString &logName);
 
 /// Returns a single custom log value
-double getSingleWorkspaceLogValue(int wsIndex,
+double getSingleWorkspaceLogValue(size_t wsIndex,
                                   const std::set<double> &logValues);
 
 #endif // MANTIDPLOTUTILITIES_H_
